Setup tsdx
We're using tsdx[1] to setup the boiler plate for the package. We've gone
with tsdx due to

a) it's used by the formik package so is proven to work well with React
package development
b) has a very sensible set of default packages
c) avoids us needing to worry about packaging and builds
d) is actively under development

More controversially this means that we're using typescript for this package.
The most compelling reason for using this here is due to the fact that the
majority of this package is going to be communicating with the prosemirror
packages, being able to ensure that we're meeting the contract set out by the
various prosemirror elements should reduce the number of bugs throughout the
package.

Given that we're going to be using cypress to handle testing of the package,
this will also help us avoid the need for ridiculous amounts of unit testing.

1) https://github.com/jaredpalmer/tsdx