Setup cypress for testing
As much as it would be great to be able to use Jest for testing, given that
jest uses jsdom as a browser we'd be unable to test the text editor without
stubbing out most of the functionality (due to jsdom not implimenting a lot
of the native browser apis relied on by prosemirror).

Instead we are using cypress, which offers us integration tests with a
similar syntax to capyabra running in a real browser. Because of this, in a
future commit we will need to setup a dummy app running the editor so that we
can visit it in the browser.