Setup example app
We need this to be able to run our cypress tests against something. This
example app uses the default setup provided by tsdx (so is using parcel to
handle everything for us).

As we get further in to the package development we'll add more and more to
the dummy app.