Setup markdown handling for prose-mirror
ProseMirror provides a reasonable set of markdown handling features[1] which
are the basis for this.

One of the specific needs for our package is the ability to completely turn
off elements (i.e. images, headings etc...).

This makes it so that our markdown handling can accept an array of nodes or
marks (think of them as block and inline elements) which we can disable.

By disabling them at the schema level we can completely turn them off (rather
than for instance needing to have special handling for paste events).

One other quirk here is that we need to make our code blocks use the fenced
code block syntax (i.e. ~~~ instead of ```). This is due to the fact that in
the rails app we use kramdown which doesn't support the backticks syntax for
codeblocks.

1) https://github.com/ProseMirror/prosemirror-markdown