Setup basic prosemirror editor
This commit introduces a function which we can use to instanciate the
prosemirror editor.

It's probably as vanilla a setup as we could hope for with basic key handling
and the use of the schema introduced in the previous commit.

The one special part here is `dispatchTransaction` which we use to fire a
callback everytime the value of the editor changes. We will use this to keep
our React component in sync with the editor.