Handle pasting into the editor
This introduces our first plugin to the editor. Plugins are ways for us to
enhance the functionality of the editor by tying in directly to prosemirror.

In this case we are handling pasting by defining the `clipboardTextParser`
and `transformPastedHTML` props.

In the case of text being pasted, we want to ensure that it's converted to
markdown using our parser so that we can ensure that the editor only contains
markdown that we support.

With html being pasted, we don't want to allow images to be pasted since we
only want to serve images that our hosted on ugc. This makes it so that if an
image is pasted we raise an error informing the user that images aren't
allowed.

Due to the fact that cypress doesn't support pasting, we've introduced a
paste command that we can use to simulate paste events with cypress.