Setup toolbar for the editor
This commit setups a toolbar (which is currently empty) that users can
interact with to toggle the formatting in the text editor.

We've introduced a `menu` plugin to handle this which ties into the `update`
lifecyle of prosemirror which means it will be fired any time the value of
prosemirror changes.

When this value changes, we check to see what marks / nodes are currently
active for the selection and fire off a state change in the react component
which updates the active items in the toolbar.

We've added a couple of utilities which can be used to help prosemirror
figure out what the active nodes and marks are within the editor.