Add support for bold text
This commit adds support for bold text to our editor.

In order to make it trivial to add new functionality to the editor we've
introduced a `BaseType` and `Mark` class which sets up some core
functionality to let the editor know whether to include it in the toolbar,
whether it's active in the schema and what the associated plugins are.

What this means in practise is that when we define a new mark, we create a
class that extends `Mark` and define it's properties within it.

For instance, if you want it to have keyboard shortcuts define them within
`shortcuts`.
For `rules` this defines the markdown that a user should type to
produce the specified mark.
`command` is what should happen when the icon is clicked in the toolbar.
`name` is the matching value in our schema.
`icon` is the name of the React component to use in the toolbar.

By defining this interface, we can infer what plugins we need to enable and
what items need to be included in the toolbar.

Note that new icons should be a react component containing the svg (we
decided not to add in any of the fancy convert svg to react handling that we
have in the main app).

For testing, any time we introduce a new mark or node, we should introduce a
corresponding integration test that tests the markdown shortcut, the keyboard
shortcut (if applicable) and the toolbar (if applicable).

With this commit, we now have a rich text editor :)