Add support for inline code
This commit follows the pattern setup in the previous mark commits.

However since this is the first instance of something that we want to disable
within the editor, this introduces a new integration spec specifically
testing that a feature can be disabled.

Due to the fact that cypress needs to visit a page rather than mounting a
component directly, we've setup an alternative page which has a component
with various features disabled which we can use to test this functionality.