Add support for headings
This commit adds support for the heading node. Even though we are supporting
all levels of headings (i.e. h1 through h6), we intend to always render them
identically. Therefore we avoid the need for the user to need to specify
which level of heading they would like via the toolbar.

One quirck of headings by default is that if a heading node has no text,
prosemirror will still flag it as a heading. We've added special handling
that says if a heading node is empty and backspace is pressed go and turn the
heading back into a paragraph.