Add support for bullet list
This follows the pattern introduced for previous nodes.

Prosemirror provides a special plugin (prosemirror-schema-list) for handling
special behaviour associated with lists (i.e. what happens when I push enter
while in a list item).

We've included this so that hopefully, the list behaviour is as predictable as possible to the end user.