Support code blocks with syntax highlighting
The first part of this follows the standard process for introducing a new
type of node.

Then things get a bit weird. We want to support automatic syntax highlighting
of code blocks to make the feature useful to learners.

To accomplish this we have written a plugin that uses lowlight[1] to
automatically highlight the text.

This works by

1. Registering the languages that we want the editor to know about (to start
with this is python, java, r, css and javascript which was determined by the
content team).

2. We then call out to `low.highlightAuto` which given a string of text will
return a value which we can use to decorate as well as a relevance score

3. When the relevance score (which is how confident low is of the language)
is greater than 1, we go through the result and decorate the content of the
code block with the results from `low.highlightAuto`.

4. We use `setNodeMarkup` to set the language as an attribute of the
codeblock which will allow rouge and kramdown to replicate the syntax
highlighting on the Rails side.

One thing to note is that our `import low from 'lowlight/lib/core';` is not
using the main entry to avoid importing all 128 languages that lowlight
supports.

1) https://github.com/wooorm/lowlight