Support image uploads
This commit makes it so that it is now possible to insert an image via drag
and drop or through a standard file upload.

When an image is uploaded we

1. Insert a placeholder into the text editor to inform the user that
something is happening

2. Make a post request to upload the image (this is since we want all images
to be hosted on ugc and go through our image processing pipeline)

3. When the request completes we replace the placeholder with the uploaded
image.

The response from the server must be of the format

```
{
  data: {
    url: 'uploadedimage.jpeg'
  }
}
```

We are using cypresses `route` feature to stub all the requests during the
specs. Due to the fact that cypress cannot stub fetch requests, we have used
an `XMLHttpRequest` so that we can test the feature properly.