Have Jenkinsfile handle releasing of the package
When someone pushes to master we want to automatically handle building the
package for use in the main app.

This makes it so that when the test suite is green on master, we build the
package to `dist`. We then use `npm version` to create a tag for the release
before pushing to master.