Add support for links
The final one of the marks that we intend to support.

We've decided not to include this in the toolbar due to the fact that we were
unable to come up with a neat ux solution for allowing users to enter the
link title and href.

Instead we are supporting links that are entered directly as well as links
that are entered as markdown (i.e. []()).

To remove it from the toolbar we include the `inToolbar` and set it to false
which will automatically exclude it from the toolbar.