Add support for blockquotes
This is our first node that we're supporting (which you can think of as a
block level element).

We've introduced a new `Node` class which extends from the `BaseType`
which sets up some node specific values, and this is what we will extend
for all node support within the editor.

To exit a blockquote a double enter will revert it to a p tag.