Use types over interfaces
Intellisense is unable to expand interface definitions so to get the
advantage of all the typescript goodness we are instead using types.

Given that we are not planning on extending these types, the recommended
approach is to write them as a type anyway so this gets us closer to best
practise.