Support multiple highlight plugins in the page
Previously the highlight plugin would only work if there was 1 active editor
on the page. This was due to the fact that we were referencing the plugin via
`highlight$`. When a second editor is on the page, the name of the plugin
then becomes `highlight$2`.

This commit changes it so that we are now using the key to access the plugin
(which is the recommended approach anyway).