Handle exiting a code block correctly
Since we added the ability to support attrs on code blocks (specifically the
language), the `isActive` check was always returning false since we weren't
passing through the attrs associated with the code block.

We don't actually care about the attrs when looking for a codeblock since we
want to toggle the selected code block regardless. This removes it so that
`nodeIsActive` now returns the correct values.

We've also removed the special handling for when toggling code blocks so that
we now just toggle the type instead of trying to do anything fancy.