Don't force image elements to behave as block level
This was leading to strange behaviour when pressing up on the cursor when the
image was selected (it would lead to the page scrolling to the top).

Looking through various prosemirror discussions it appears that this might be
related to a browser bug.

Rather than trying to fix chrome, I've decided to stop trying to pretend that
images are block level elements. Although this might be slightly strange
behaviour for the end user (i.e. you can insert text inline with an image),
this is better than the page scrolling away from the text area.