Ignore attrs when checking if a nodeIsActive
Since we set the attrs to reflect the language of a code_block, our check for
whether a code_block was active or not was always returning false if the
code_block was syntax highlighted.

We don't care about whether the attrs match or not since we know if we've
find a code_block that it is active in the selection (regardless of it's
attrs).

We are now just returning true or false based on whether node exists or not.