Hide Toolbar & ContextualHelp
This commit uses the editable prop to decide whether to render the
Toolbar and ContextualHelp in the editor. When in not-editable mode
we don't want to show the user the toolbar. If the content
being rendered is a code block the ContextualHelp was still firing
even in a non-editable context, so I've hidden that too.