Pass editable prop into the Editor plugin
This PR aims to enable the functionality in promisemirror to make the
text editor readonly (their terminology is 'not editable'). It sets
a default editable prop value of 'true' so it can be overridden only in
the  scenario where we want a not editable version of the editor

The editable prop expects a function which I've not tied to the
`EditorState` object because we're not looking for its editable state
to toggle only render as completely not-editable or not.

In future commits I'll also look to disable the toolbar & contextual
help text that appears when the user is writing a code block, as well
as adding some integration tests

[1] https://prosemirror.net/docs/ref/#view.EditorProps.editable