Add integration tests for not editable text Editor
This commit adds some integration tests for the text editor in a not
editable context. I've opted to add a new example view for this set of
tests

I'd prefer if the test that a user can't type in the editor was more
behavioural but I couldn't find a way to use cypress tools to do this
so it hooks more directly into the prosemirror implementation detail of
the contenteditable attribute

Also the final test around contextual help - the multiple clicks are in
place to avoid a false positive where fewer clicks allowed the
expectation to pass without changing the code to actually hide the help.
I spent some time experimenting with waits & other cypress tools around
interacting with the DOM but couldn't get a failure without a couple of
extra clicks. These clicks run faster than any waiting so felt it was
a reasonable tradeoff.