Specify the node version more comprehensively
GitHub doesn't appear to support reading the version from the files I've
added, so I've retained that and changed it to match.

These two files should catch most automatic node version switcher tools
so that engineers are running the right node version.

I chose 10.24.1 as it's the most recent 10.x.x version.

This is well below what we use in the main app (14.5.5) but upgrading is
a separate job.