Ignore typescript errors in Plugins/highlight.ts
We want to upgrade to Node V14, but before we do that I want to ensure
that we are using the latest version of our build tools (for example
tsdx and typescript) in order to avoid incompatibility issues.

I noticed a couple of Typescript linting errors in our highlight plugin
and thought it was worth fixing those before upgrading Typescript.