Bump rollup-plugin-postcss 4.0.1
Before we upgrade Node I want to first ensure that we are using the
latest version of our build dependencies in order to reduce errors due
to compatibility issues.

One of the key changes with `rollup-plugin-postcss` is that `postcss` is
now a peer dependency and so we now neeed have it as an explicit
devDependency.

I have also upgraded `node-sass` to ensure that these libraries are
compatible, but I will be switching from `node-sass` to `sass` in a
later commit as `node-sass` is now deprecated.