Bump Node V14
https://trello.com/c/KpSklxda

The main motivation for this upgrade is so that we can upgrade to Parcel
v2 which doesn't work on Node v10.

It is worth noting Parcel V1 doesn't run on Node v14 and so this commit
will break the build, however I thought it was useful to split this out
the upgrades into separate commits. Parcel will be upgraded in
subsequent commits.