Fix @rollup/plugin-replace warning
After upgrading TSDX and @rollup/plugin-replace I am now seeing a
warning when I run `npm run build`:

@rollup/plugin-replace: 'preventAssignment' currently defaults to false.
It is recommended to set this option to `true`, as the next major
version will default this option to `true`.

After some Googling the suggested solution [1] for the time being is to
override the options passed into the plugin in our `tsdx.config.js`
file.

[1] https://github.com/jaredpalmer/tsdx/issues/981