Use postcss-import plugin for inlining sass imports
I want to switch out node-sass for sass since node-sass is now
deprecated [1].

Previously we were relying on the `rollup-plugin-postcss` plugin for
inlining our sass imports, however switching from `node-sass` to `sass`
appears to break this functionality. I have been unable to find a
related github issue, but using the postcss-import plugin appears to
resolve this issue and feels like the simplest solution. As we are now
relying on postcss-import for this functionality I have removed the `~`
prefixes.

[1] https://sass-lang.com/blog/libsass-is-deprecated